Added Abstraction
Each sort class now inherits from the new Sort superclass which gives acces to an isSorted() method and requires the implementation of a sort() method. The sorting algorithm methods no longer return the sorted array, they instead output to the command line whether the sort was successful or not. README updated.